Update README.md
Just an update to the readme file in the readme-edits branch of the master branch. Next I will commit this to the repository and submit a pull request for the changes to the README file. This is used to start a review of the changes before being merged with the branch.